+++ .env_example
@@ -2,7 +2,7 @@ STEAM_API_KEY=REPLACE_WITH_KEY
 SESSION_SECRET=somesecretvalue
 ROOT_URL=http://yourdomain.com
 START_SEQ_NUM=AUTO
-RETRIEVER_HOST=http://retrieverhost.com,http://retrieverhost2.com
+RETRIEVER_HOST=http://localhost:5300,http://retrieverhost2.com
 RECAPTCHA_PUBLIC_KEY=REPLACE_WITH_KEY
 RECAPTCHA_SECRET_KEY=REPLACE_WITH_KEY
 KUE_USER=user
